Refactor dyn into private module
